database/requirements.txt
psycopg2-binary==2.9.9
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pyjwt==2.8.0
tabulate==0.9.0
gunicorn==21.2.0
postgis==1.0.4
geoalchemy2==0.14.0
sqlalchemy==2.0.32
pandas==2.2.0 